Refactor: 새로운 Todo 입력 기능 컴포넌트화하여 적용 #6
